Added Comments in the Fragmenter Class

+++ GraphMaker/src/VisualizeBinary/Fragmenter.java
@@ -14,8 +14,9 @@ import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 
 /**
+ * This class will take a single file, multiple files or an entire directory and fragment it into 512 byte fragments.
+ * If a file is less than the 512 byte section, it fills the end in with zeros. (not sure if proper way to do this...)
  * @author hartell
- *
  */
 public class Fragmenter {
 
@@ -29,68 +30,79 @@ public class Fragmenter {
 		
 		//Create a new file chooser based for either directories or files.
 		JFileChooser chooser = new JFileChooser();
+		//Starts the program in a specific directory
 		chooser.setCurrentDirectory(new File("H://SVM/"));
+		//If the user wants to fragment a directory...
 		if(answer == 0){
 			System.out.println("User selected a fragment a Directory");
 			chooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
 		}
+		//Otherwise fragment the files they selected...
 		else{
 			System.out.println("User selected to fragment File/s");
 			chooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
 			chooser.setMultiSelectionEnabled(true);
 		}
 		
-		//Let the user select what they want to fragment
+		//Ask the user to select a file/s or a directory.
 		int selection = chooser.showOpenDialog(null);
 		
+		//Ok they've selected something, get what they've selected.
 		if(selection == JFileChooser.APPROVE_OPTION){
 			File[] files = null;
-			//aka file/s were selected
+			//If file/s were selected...
 			if(answer == 1){ 
 				files = chooser.getSelectedFiles();
 			}
-			//it's a directory so grab all the files inside of the directory
+			//it's a directory so grab all ONLY files (no directories) inside of the directory
 			else {
 				File dir = chooser.getSelectedFile();
 				System.out.println("Dir Path: " + dir.getAbsolutePath());
 				if(dir.isDirectory()){
+					//Get a list of all the files/directories inside the directory
 					File[] listOfFiles = dir.listFiles();
 					ArrayList<File> tempFileHolder = new ArrayList<File>();
+					//Remove all directories
 					for(File f : listOfFiles){
 						if(!f.isDirectory()){
 							tempFileHolder.add(f);
 						}
 					}
-					
+					//Add the just the files (no directories) into the list of selected files.
 					files = new File[tempFileHolder.size()];
 					for(int i = 0; i < tempFileHolder.size(); i++){
 						files[i] = tempFileHolder.get(i);
 					}
-					System.out.println(files.length);
 				}
-				else
+				else{
+					//Otherwise something went wrong...
 					System.out.println("Directory not selected");
+				}
 			}
 			
-			//Create a Directory to put the file fragments in if it's not already there...
+			//Create a Directory to put the file fragments
 			File output = new File(files[0].getParentFile() + "\\Fragmented");
 			
-			// create a directory to put file fragments in if its not already there
+			//Check to see if the directory exists...
+			//If not, make it.
 			if(!output.isDirectory()){
 				output.mkdirs();
 			} 
+			//If it does exist, then ask the user if they want to delete current fragments or append to the current fragments
 			else if(output.isDirectory()){
 				//Ask user if they want to delete currently existing fragments
 				int emptyFolder = JOptionPane.showConfirmDialog(null, "Delete current Fragments?", "Delete?", JOptionPane.YES_NO_OPTION);
 				
-				//TODO: Delete files in Directory
+				//Delete files in Directory
 				if(emptyFolder == JOptionPane.YES_OPTION){
-					System.out.println("Deleting Old Fragments");
+					System.out.println("Deleting Old Fragments...");
 					//Empty the folder
 					File[] fileList = output.listFiles();
 					for(File myFile : fileList){
-						System.out.println("Trying to delete: " + myFile.getAbsolutePath());
-						myFile.delete();
+						//Delete only non-directories...
+						if(!myFile.isDirectory()){
+							myFile.delete();
+						}
 					}
 				}
 				else{
@@ -98,9 +110,11 @@ public class Fragmenter {
 				}
 			}
 			else {
+				//Something went wrong
 				throw new IllegalArgumentException("Output directory is not a directory!");
 			}
 			
+			//Iterate through the files the user wants to fragment, name them fragment + (fragCounter - aka wherever the last filename left off);
 			int i = 1;
 			int fragCounter = 1;
 			for(File f : files){
@@ -112,26 +126,40 @@ public class Fragmenter {
 				i++;
 			}
 		}
+		System.out.println("Fragmentation Complete");
 	}
 	
-	public static int fragmentFile(File input, File output, int fragCounter){	
+	/**
+	 * This method fragments a file into 512 bytes sections. Remnants of files < 512 bytes are padded with zeros (not sure if right...)
+	 * @param input - the file to be fragmented
+	 * @param output - the output location
+	 * @param fragCounter - the last numerical fragment number
+	 * @return int - value of the last file created.
+	 */
+	public static int fragmentFile(File input, File output, int fragCounter){
+		//Thanks to Ben Holland for assistance with RandomAccessFiles!
 		int fileCounter = fragCounter;
-		//Create a randomAccessFile to fragment from
 		try {
+			//Create a randomAccessFile to fragment from
 			RandomAccessFile file = new RandomAccessFile(input, "r");
-			
+			//While the file has more than 512 bytes left...
 			while(file.getFilePointer()+512 < file.length()){
+				//Read in 512 bytes
 				byte[] buffer = new byte[512];
 				file.readFully(buffer);
+				//Output the 512 byte fragment to a incremented file named Fragment + fileCount
 				OutputStream out = new FileOutputStream(new File(output.getAbsolutePath() + File.separatorChar + "Fragmented" + fileCounter));
 				fileCounter++;
 				out.write(buffer);
 				out.close();
 			}
+			//IF the file has any leftover bytes < 512 left...
 			if(file.getFilePointer() != file.length()-1){
+				//Read in whatever is left and pad with 0s upto 512 bytes.
 				byte[] buffer = new byte[512];
 				int end = (int) (file.length() - 1 - file.getFilePointer());
 				file.readFully(buffer, 0 , end);
+				//Output the 512 byte fragment to a incremented file named Fragment + fileCount
 				OutputStream out = new FileOutputStream(new File(output.getAbsolutePath() + File.separatorChar + "Fragmented" + fileCounter));
 				fileCounter++;
 				out.write(buffer);
